Fix imports in tests to account for refactor
